sdk/python/requirements.txt
python-jose[cryptography]>=3.3.0
aiohttp>=3.8.0
